// frontend/src/api/models/listPurchaseReturnsStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListPurchaseReturnsStatus =
  (typeof ListPurchaseReturnsStatus)[keyof typeof ListPurchaseReturnsStatus]

export const ListPurchaseReturnsStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  SHIPPED: 'SHIPPED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const