// frontend/src/api/models/handlerSupplierListResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerSupplierListResponseStatus } from './handlerSupplierListResponseStatus'

/**
 * Supplier list item with basic information
 */
export interface HandlerSupplierListResponse {
  city?: string
  code?: string
  created_at?: string
  email?: string
  id?: string
  name?: string
  payment_term_days?: number
  phone?: string
  province?: string
  rating?: number
  short_name?: string
  sort_order?: number
  status?: HandlerSupplierListResponseStatus
}